replaces zeros with NaNs when importing data; change kymograph palette
**Data import**

The first step of the pipeline is importing data.  Fluorescence reading
data is usually in a CSV file with specific headers, and
pipeline/dataimport.py is written to read and parse such a file.

Time points with no cell are represented as a '0'.  This causes a
semantic problem i.e. is it no cell or a cell with no fluorescence?  In
addition, when computing the mean of a time trace, the values will be
skewed -- previously this wasn't a problem as I exclusively set the
'remain' flag to be 1.0.  Replacing the zeros with NaNs before the df is
returned is thus the most logical way, but further use of the data must
include proper handling of NaN values e.g. using numpy.nanmean() instead
of numpy.mean().

**Kymograph palette**

Following Peter's suggestion: new palette is more vivid and will colour
oscillation troughs blue and oscillation peaks red.  Previously it was
just a one-direction gradient, thus difficult to see oscillations
clearly.