remove pipeline's dependence on the specifics of certain experiments
The main OscillationClassifier.py script contains a sample use of the
whole pipeline, from data import, processing into objects, time series
processing, and periodogram computations.

I have made derivations specific to each experiment
(OscillationClassifier*.py), but the issue is... they are specific to
each experiment.  This script attempts to create a more general case.
Specifically,
- It uses just one DatasetAttr object, Dset, rather than one for each
dataset.  The intention is that the script is run once for each dataset
(in most cases); this does not preclude using it with multiple datasets,
but I think this paradigm is best for compartmentalising workflow.
- It requires a *strains.csv file that specifies the strain for each
position.  Previously, I had to manually define the strain for each
position; this file provides a nice look-up table.
- It implements a try-except structure so that it can be used with both
experiments that employ only flavin and experiments that use both flavin
and mCherry (though the strain has to be 'htb2_mCherry_CRISPR', but it
is the only mCherry-containing strain I actively use).

Furthermore, it provides options to:
- Subset time series to a specifc range of time points; for example,
during a specific pump regime.
- Normalise time series (in *.reading_processed) so that time points are
expressed as deviations from the mean, relative to the range.

These options are meant to work with the new palette used by the
kymograph (commit 38f97f2).

I made these changes in order to generate plots for my 2nd-year review,
but they will also be useful in future when I re-factor this pipeline to
work with swain-lab/python-pipeline/post-processing.