Format data file from experiment 19972 for BioDare
BioDare (https://biodare2.ed.ac.uk/; Zielinski et al. 2014
doi:10.1371/journal.pone.0096462) offers an on-line repository and
algorithms to clean-up and perform period analysis of time series data.
This service requires data to be formatted in a way that rows are time
points and columns are cells; the first column contains a list of time
points.  This is different from how my data is formatted, so I need to
re-format data files to test the BioDare service.

Note: I've tried and it fails to do anything useful so far -- but not
ruling out further follow-up.