Changed format of birth time data spreadsheets
aliby produces birth time dataframes; their format is a binary matrix
(elements are 0 or 1).  Currently, the birth data is birth time in
minutes.  For cycle alignment, I plan to do things as close to how
things are in aliby as possible, plus the birth times from the existing
data may not be accurate (i.e. they are not divisible by 5 for some
reason).  Hence the changed format of data.