Enable subsetting catch22 features
Reduction of features -- focusing on the ones that are most important --
can be useful.  In addition,
more information doesn't necessarily mean a better classifier, as the
model will try to do something with all the features and may introduce
bias.  Some catch22 features may be noisy, fully explainable by other
features, or correlated to another feature.