Refactor: mutual information uses aliby-style import/processing/featurisation
This makes code cleaner and matches svm, cycle_alignment, and umap
notebooks.
(One of these days I'll replace the import steps with importing HDF5
files.  Then put the processing & featurisation in a separate script
that outputs a CSV that can be re-imported, so that these notebooks each
do one thing.)