Produce binary matrix of peaks of flavin cycles
Added cells that define a function to produce a binary matrix of peaks
of flavin cycles -- i.e., a wrapper for scipy.find_peaks() -- and apply
this function to the signal_flavin_processed DataFrame.

Did it this way so that I have a function that nicely works with
pandas.DataFrame.apply().

Also:
* Re-ordered cells.  It makes senes to have births and flavin peaks at
  the same 'level' of 'hierarchy' since they both involve binary
  matrices.  And I plan to do similar alignment & visualisation tasks.
* Added some blank cells for future histograms.

To do: manage variable names -- births and flavin masks must have
different variable names.